last bug fixes for push to main
