+++ DESCRIPTION
@@ -66,7 +66,7 @@ Imports:
     lpSolve,
     WienR
 LinkingTo: 
-    Rcpp, RcppArmadillo, RcppEigen
+    Rcpp, RcppArmadillo
 Depends: 
     R (>= 3.5.0)
 LazyData: true
