+++ R/RcppExports.R
@@ -1,20 +1,12 @@
 # Generated by using Rcpp::compileAttributes() -> do not edit by hand
 # Generator token: 10BE3573-1514-4C36-9D1C-5A225CD40393
 
-calculate_subject_means <- function(group_designs, params, n_subjects, n_pars) {
-    .Call(`_EMC2_calculate_subject_means`, group_designs, params, n_subjects, n_pars)
-}
-
-standard_subj_ll <- function(group_designs, theta_var, theta_mu, alpha, n_subj) {
-    .Call(`_EMC2_standard_subj_ll`, group_designs, theta_var, theta_mu, alpha, n_subj)
-}
-
-dmvnorm_cpp <- function(X, mean, sigma, logd = FALSE) {
-    .Call(`_EMC2_dmvnorm_cpp`, X, mean, sigma, logd)
-}
-
 sp_new <- function(iter, lambda_varimax, q, p, dim_all_c, all_c, lambda_hat, st, cost_matrix, perm) {
     .Call(`_EMC2_sp_new`, iter, lambda_varimax, q, p, dim_all_c, all_c, lambda_hat, st, cost_matrix, perm)
+}
+
+calculate_subject_means <- function(group_designs, params, n_subjects, n_pars) {
+    .Call(`_EMC2_calculate_subject_means`, group_designs, params, n_subjects, n_pars)
 }
 
 dlba <- function(t, A, b, v, sv, posdrift = TRUE) {
