+++ R/variant_standard.R
@@ -501,7 +501,7 @@ bridge_group_and_prior_and_jac_standard <- function(
   prior_mu_log <- dmvnorm(theta_mu,
                           mean  = prior$theta_mu_mean,
                           sigma = prior$theta_mu_var,
-                          logd   = TRUE)
+                          log   = TRUE)
 
   # For the Jacobian of var1 => rowSums(theta_var1)
   # For the Jacobian of a    => rowSums(theta_a)
@@ -560,7 +560,7 @@ bridge_group_and_prior_and_jac_standard <- function(
         par_idx <- par_idx + ncol(group_designs[[k]])
       }
       # Now alpha_s ~ N(mu_s, var_curr)
-      group_ll <- group_ll + dmvnorm(alpha_s, mu_s, var_curr, logd = TRUE)
+      group_ll <- group_ll + dmvnorm(alpha_s, mu_s, var_curr, log = TRUE)
     }
 
     # 4) Prior on var1, var2, and a => same partial-block logic
@@ -646,14 +646,6 @@ bridge_add_group_standard <- function(all_samples, samples, idx){
   return(all_samples)
 }
 
-dmvnorm <- function(x, mean, sigma, logd = FALSE)
-{
-  if (is.null(dim(x)))            # plain numeric vector?
-    x <- matrix(x, nrow = 1)      # → 1 × p
-
-  dmvnorm_cpp(x, mean, sigma, logd)
-}
-
 # for IC ------------------------------------------------------------------
 
 group__IC_standard <- function(emc, stage="sample", filter=NULL) {
@@ -677,9 +669,9 @@ group__IC_standard <- function(emc, stage="sample", filter=NULL) {
   if(is.null(emc[[1]]$group_designs)){
     lls <- numeric(N)
     for(i in 1:N){
-      lls[i] <- sum(dmvnorm(t(alpha[,,i]), theta_mu[,i], theta_var[,,i], logd = T))
+      lls[i] <- sum(dmvnorm(t(alpha[,,i]), theta_mu[,i], theta_var[,,i], log = T))
     }
-    mean_pars_ll <-  sum(dmvnorm(t(mean_alpha), mean_mu, mean_var, logd = TRUE))
+    mean_pars_ll <-  sum(dmvnorm(t(mean_alpha), mean_mu, mean_var, log = TRUE))
   } else{
     # 3) Build design matrices
     group_designs <- add_group_design(emc[[1]]$par_names, emc[[1]]$group_designs, n_subj)
@@ -691,7 +683,7 @@ group__IC_standard <- function(emc, stage="sample", filter=NULL) {
     mean_pars_ll <- 0
     for(s in seq_len(n_subj)) {
       mean_pars_ll <- mean_pars_ll +
-        dmvnorm(mean_alpha[, s], mean_subj_means[, s], mean_var, logd=TRUE)
+        dmvnorm(mean_alpha[, s], mean_subj_means[, s], mean_var, log=TRUE)
     }
 
   }
@@ -700,4 +692,26 @@ group__IC_standard <- function(emc, stage="sample", filter=NULL) {
   mean_ll <- mean(lls)
   Dmean <- -2*mean_pars_ll
   list(mean_ll = mean_ll, Dmean = Dmean, minD = minD)
+}
+
+standard_subj_ll <- function(theta_var, theta_mu, alpha, n_subj, group_designs)
+{
+  N <- dim(theta_var)[3];  p <- nrow(theta_mu)
+  log2pi <- log(2*pi)
+
+  # pre‑allocate
+  ll <- numeric(N)
+
+  for (i in seq_len(N)) {
+
+    Sigma <- theta_var[,,i]
+    U     <- chol(Sigma)              # will error if non‑PD
+    rooti <- backsolve(U, diag(p))
+    log_const <- sum(log(diag(rooti))) - 0.5 * p * log2pi
+
+    mu  <- calculate_subject_means(group_designs, theta_mu[,i], n_subj, p)
+    z   <- t(alpha[,,i] - mu) %*% rooti      # n_subj × p
+    ll[i] <- -0.5 * sum(z^2) + n_subj*log_const
+  }
+  ll
 }
