+++ src/RcppExports.cpp
@@ -2,7 +2,6 @@
 // Generator token: 10BE3573-1514-4C36-9D1C-5A225CD40393
 
 #include <RcppArmadillo.h>
-#include <RcppEigen.h>
 #include <Rcpp.h>
 
 using namespace Rcpp;
@@ -12,49 +11,6 @@ Rcpp::Rostream<true>&  Rcpp::Rcout = Rcpp::Rcpp_cout_get();
 Rcpp::Rostream<false>& Rcpp::Rcerr = Rcpp::Rcpp_cerr_get();
 #endif
 
-// calculate_subject_means
-arma::mat calculate_subject_means(const Rcpp::List& group_designs, const arma::colvec& params, const int n_subjects, const int n_pars);
-RcppExport SEXP _EMC2_calculate_subject_means(SEXP group_designsSEXP, SEXP paramsSEXP, SEXP n_subjectsSEXP, SEXP n_parsSEXP) {
-BEGIN_RCPP
-    Rcpp::RObject rcpp_result_gen;
-    Rcpp::RNGScope rcpp_rngScope_gen;
-    Rcpp::traits::input_parameter< const Rcpp::List& >::type group_designs(group_designsSEXP);
-    Rcpp::traits::input_parameter< const arma::colvec& >::type params(paramsSEXP);
-    Rcpp::traits::input_parameter< const int >::type n_subjects(n_subjectsSEXP);
-    Rcpp::traits::input_parameter< const int >::type n_pars(n_parsSEXP);
-    rcpp_result_gen = Rcpp::wrap(calculate_subject_means(group_designs, params, n_subjects, n_pars));
-    return rcpp_result_gen;
-END_RCPP
-}
-// standard_subj_ll
-arma::vec standard_subj_ll(const Rcpp::List& group_designs, const arma::cube& theta_var, const arma::mat& theta_mu, const arma::cube& alpha, const int n_subj);
-RcppExport SEXP _EMC2_standard_subj_ll(SEXP group_designsSEXP, SEXP theta_varSEXP, SEXP theta_muSEXP, SEXP alphaSEXP, SEXP n_subjSEXP) {
-BEGIN_RCPP
-    Rcpp::RObject rcpp_result_gen;
-    Rcpp::RNGScope rcpp_rngScope_gen;
-    Rcpp::traits::input_parameter< const Rcpp::List& >::type group_designs(group_designsSEXP);
-    Rcpp::traits::input_parameter< const arma::cube& >::type theta_var(theta_varSEXP);
-    Rcpp::traits::input_parameter< const arma::mat& >::type theta_mu(theta_muSEXP);
-    Rcpp::traits::input_parameter< const arma::cube& >::type alpha(alphaSEXP);
-    Rcpp::traits::input_parameter< const int >::type n_subj(n_subjSEXP);
-    rcpp_result_gen = Rcpp::wrap(standard_subj_ll(group_designs, theta_var, theta_mu, alpha, n_subj));
-    return rcpp_result_gen;
-END_RCPP
-}
-// dmvnorm_cpp
-Eigen::VectorXd dmvnorm_cpp(const Eigen::MatrixXd& X, const Eigen::RowVectorXd mean, const Eigen::MatrixXd& sigma, const bool logd);
-RcppExport SEXP _EMC2_dmvnorm_cpp(SEXP XSEXP, SEXP meanSEXP, SEXP sigmaSEXP, SEXP logdSEXP) {
-BEGIN_RCPP
-    Rcpp::RObject rcpp_result_gen;
-    Rcpp::RNGScope rcpp_rngScope_gen;
-    Rcpp::traits::input_parameter< const Eigen::MatrixXd& >::type X(XSEXP);
-    Rcpp::traits::input_parameter< const Eigen::RowVectorXd >::type mean(meanSEXP);
-    Rcpp::traits::input_parameter< const Eigen::MatrixXd& >::type sigma(sigmaSEXP);
-    Rcpp::traits::input_parameter< const bool >::type logd(logdSEXP);
-    rcpp_result_gen = Rcpp::wrap(dmvnorm_cpp(X, mean, sigma, logd));
-    return rcpp_result_gen;
-END_RCPP
-}
 // sp_new
 Rcpp::List sp_new(const int iter, const arma::cube& lambda_varimax, const int q, const int p, const int dim_all_c, const arma::mat& all_c, const arma::mat& lambda_hat, const arma::uvec& st, arma::mat cost_matrix, arma::mat perm);
 RcppExport SEXP _EMC2_sp_new(SEXP iterSEXP, SEXP lambda_varimaxSEXP, SEXP qSEXP, SEXP pSEXP, SEXP dim_all_cSEXP, SEXP all_cSEXP, SEXP lambda_hatSEXP, SEXP stSEXP, SEXP cost_matrixSEXP, SEXP permSEXP) {
@@ -72,6 +28,20 @@ BEGIN_RCPP
     Rcpp::traits::input_parameter< arma::mat >::type cost_matrix(cost_matrixSEXP);
     Rcpp::traits::input_parameter< arma::mat >::type perm(permSEXP);
     rcpp_result_gen = Rcpp::wrap(sp_new(iter, lambda_varimax, q, p, dim_all_c, all_c, lambda_hat, st, cost_matrix, perm));
+    return rcpp_result_gen;
+END_RCPP
+}
+// calculate_subject_means
+arma::mat calculate_subject_means(const Rcpp::List& group_designs, const arma::colvec& params, const int n_subjects, const int n_pars);
+RcppExport SEXP _EMC2_calculate_subject_means(SEXP group_designsSEXP, SEXP paramsSEXP, SEXP n_subjectsSEXP, SEXP n_parsSEXP) {
+BEGIN_RCPP
+    Rcpp::RObject rcpp_result_gen;
+    Rcpp::RNGScope rcpp_rngScope_gen;
+    Rcpp::traits::input_parameter< const Rcpp::List& >::type group_designs(group_designsSEXP);
+    Rcpp::traits::input_parameter< const arma::colvec& >::type params(paramsSEXP);
+    Rcpp::traits::input_parameter< const int >::type n_subjects(n_subjectsSEXP);
+    Rcpp::traits::input_parameter< const int >::type n_pars(n_parsSEXP);
+    rcpp_result_gen = Rcpp::wrap(calculate_subject_means(group_designs, params, n_subjects, n_pars));
     return rcpp_result_gen;
 END_RCPP
 }
@@ -502,10 +472,8 @@ END_RCPP
 }
 
 static const R_CallMethodDef CallEntries[] = {
-    {"_EMC2_calculate_subject_means", (DL_FUNC) &_EMC2_calculate_subject_means, 4},
-    {"_EMC2_standard_subj_ll", (DL_FUNC) &_EMC2_standard_subj_ll, 5},
-    {"_EMC2_dmvnorm_cpp", (DL_FUNC) &_EMC2_dmvnorm_cpp, 4},
     {"_EMC2_sp_new", (DL_FUNC) &_EMC2_sp_new, 10},
+    {"_EMC2_calculate_subject_means", (DL_FUNC) &_EMC2_calculate_subject_means, 4},
     {"_EMC2_dlba", (DL_FUNC) &_EMC2_dlba, 6},
     {"_EMC2_plba", (DL_FUNC) &_EMC2_plba, 6},
     {"_EMC2_dWald", (DL_FUNC) &_EMC2_dWald, 5},
